+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased](https://github.com/inspirum/balikobot-php/compare/v5.4.1...master)
+## [Unreleased](https://github.com/inspirum/balikobot-php/compare/v5.5.0...master)
+
+
+## [v5.5.0 (2021-07-21)](https://github.com/inspirum/balikobot-php/compare/v5.4.1...v5.5.0)
 ### Added
 - Added support for new options from documentation **v1.929** (2021-07-13)
   - Add **KURIER** shipper